milestone: we can compile.
refactored a bit, moved around the classes to make them.... more pleasing to the eye? easier for me to understand?

whatever, they're not all over the freakin place now.

also migrated to androidx because why the hell not...

spek may or may not be working but the implements are there and supported stuff.

oh my gosh i'm not up for this today, i want to play with my calculator app -- ok ok lets knock this out