installing retrofit seems like the best plan, why? here's my thinking:
a: it is a sphaghetti mess to sort out exactly how everything is getting downloaded and parsed and non of it is threaded, we could go through pick and pull and fix but ! I have little experience using the methods used in creating this app, this will make it take longer and make things a lot harder going than just starting over.

b: it will be faster, it will also be naturally async. this will lead to performance gains.

c: I'm already getting close to the deadline and this is a hail mary.